rename `types.d.ts` to `core.d.ts`
