+++ src/types/core.d.ts
